Backend settings saving rewrite
Holy hell this one was a pain.

* Rewrote saving system to be fully automated by looping through the
array of settings in each menu instead of going through hardcoded
values. This was a pain in the ass but will be worth it when I have to
implement a crapton of settings in the long run.
* GUI settings objects take in a string of a setting instead of the
setting itself, to allow for the aforementioned rewrite.